+++ engine/apps/slack/scenarios/slack_channel_integration.py
@@ -1,9 +1,9 @@
 import logging
 import typing
 
-from django.core.exceptions import ObjectDoesNotExist
-
+from apps.slack.client import SlackAPIException
 from apps.slack.scenarios import scenario_step
+from apps.slack.scenarios.resolution_note import handle_resolution_note_message_exception
 from apps.slack.types import EventPayload, EventType, MessageEventSubtype, PayloadType, ScenarioRoute
 
 if typing.TYPE_CHECKING:
@@ -73,24 +73,22 @@ class SlackChannelMessageEventStep(scenario_step.ScenarioStep):
         except SlackMessage.DoesNotExist:
             return
 
-        try:
-            alert_group = slack_message.get_alert_group()
-        except ObjectDoesNotExist:
+        if not slack_message.alert_group:
             # SlackMessage instances without alert_group set (e.g., SSR Slack messages)
             return
 
-        result = self._slack_client.api_call(
-            "chat.getPermalink",
-            channel=channel,
-            message_ts=message_ts,
-        )
+        try:
+            result = self._slack_client.chat_getPermalink(channel=channel, message_ts=message_ts)
+        except SlackAPIException as e:
+            handle_resolution_note_message_exception(self, "save thread message", e)
+            return
+
         permalink = None
         if result["permalink"] is not None:
             permalink = result["permalink"]
 
         if len(text) > 2900:
-            self._slack_client.api_call(
-                "chat.postEphemeral",
+            self._slack_client.chat_postEphemeral(
                 channel=channel,
                 user=slack_user_identity.slack_id,
                 text=":warning: Unable to show the <{}|message> in Resolution Note: the message is too long ({}). "
@@ -101,7 +99,7 @@ class SlackChannelMessageEventStep(scenario_step.ScenarioStep):
         slack_thread_message, created = ResolutionNoteSlackMessage.objects.get_or_create(
             ts=message_ts,
             thread_ts=thread_ts,
-            alert_group=alert_group,
+            alert_group=slack_message.alert_group,
             defaults={
                 "user": self.user,
                 "added_by_user": self.user,